Revert "esp_partition: disable encrypted reads/writes if flash encryption is disabled"
This reverts commit 0ca0d02aab8a88e8e11cbfbb6bfe49b3c62b06f5.

It has been noticed that there are scenarios where even though firmware is not enabled
with flash encryption config feature, it should be able to write to encrypted partitions.

This revert adds the feature back which was removed around v4.0 timelines, and same
change will be backported to all releases (upto v4.0) for consistency.