esp_netif: Set default netif with lwip_netif
Make use of lwip_netif pointer for setting the default interface in lwip,
since the  is a generic pointer used in I/O functions wheras the former is always the  type.
More importantly the netif_handle could be set to another context ptr
used in I/O functions, as as for the SLIP netif.

This change fixes the issue of incorrect settings of a default interface
if any SLIP netif is involved.

Closes [Github: espressif/esp-idf]/issues/7246