submodules: remove wifi and bt submodules
In order to allow Zephyr's CI to pass, wifi and bt
submodules were removed from hal_espressif as it causes
CI to fail. This PR adds a new file which contains submodules
paths and revision so it can be tracked.
Python script enables the download by using the extension
"west espressif update", which triggers the download of those
repositories.

Signed-off-by: Sylvio Alves <sylvio.alves@espressif.com>