bt: adapter: fixup bt pwr on/off
This implements a fix for powering on/off the bt domain.
This fixes an issue where when using bluetooth with the
esp32c3 on Zephyr, the application would immediately hang
after the boot banner.

Althought the fix for power down is implemented,
it is not tested as the esp32c3 `hci_driver` close
is not impl.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>