heap: add SPIRAM heap offset and size
The existing heap starts at the beginning of SPIRAM which is also occupied
by the WiFi stack if CONFIG_ESP32_WIFI_NET_ALLOC_SPIRAM is enabled.  This
results in memory corruption.

Update the heap code to use a linker-description symbol _spiram_heap_start
for setting the starting point of the SPIRAM heap.

Requires matching change to the .ld files in the Zephyr soc code.

Fixes: 61359

Signed-off-by: Eric Holmberg <eric.holmberg@northriversystems.co.nz>