esptool: Pad sections to account for alignment
Parse ELF section alignment values and use those to check to see if
successive sections are separated because of them. Pad a previous
section if doing so would cause it to precisely align with the next
section start address. This ensures that alignment requirements don't
spoil our attempt to merge sections.

Signed-off-by: Keith Packard <keithp@keithp.com>