scripts: blobs: Add non-standard rule for esp32c2
ESP32-C2 BT blobs repository doesn't have a subfolder named after
the SoC. This requires a change in the script to account for such
cases. When no subfolder with an SoC is found, the script will
match for the SoC using the name of the repository and attempt to
find library files in the root directory instead.

Signed-off-by: Raffael Rostagno <raffael.rostagno@espressif.com>